+++ lib/src/ui/dock_layout_panel.dart
@@ -75,12 +75,10 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
    *
    * @param unit the unit to be used for layout
    */
-  DockLayoutPanel(Unit unit) {
-    this._unit = unit;
-
+  DockLayoutPanel(this._unit) {
     setElement(new dart_html.DivElement());
     _layout = new Layout(getElement());
-    _layoutCmd = new DockAnimateCommand(this, _layout);
+    _layoutCmd = new _DockAnimateCommand(this, _layout);
   }
 
   //*********
@@ -94,7 +92,7 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
    * @param widget the widget to be added
    */
   void add(Widget widget) {
-    _insert(widget, LayoutDirection.CENTER, 0.0, null);
+    _insert(widget, DockLayoutConstant.CENTER, 0.0, null);
   }
 
   //*****
@@ -106,7 +104,7 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
    * @param size the child widget's size
    */
   void addEast(Widget widget, double size) {
-    _insert(widget, LayoutDirection.EAST, size, null);
+    _insert(widget, DockLayoutConstant.EAST, size, null);
   }
 
   /**
@@ -126,7 +124,7 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
    * @param size the child widget's size
    */
   void addLineEnd(Widget widget, double size) {
-    _insert(widget, LayoutDirection.LINE_END, size, null);
+    _insert(widget, DockLayoutConstant.LINE_END, size, null);
   }
 
   /**
@@ -137,7 +135,7 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
    * @param size the child widget's size
    */
   void addLineStart(Widget widget, double size) {
-    _insert(widget, LayoutDirection.LINE_START, size, null);
+    _insert(widget, DockLayoutConstant.LINE_START, size, null);
   }
 
   /**
@@ -147,7 +145,7 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
    * @param size the child widget's size
    */
   void addNorth(Widget widget, double size) {
-    _insert(widget, LayoutDirection.NORTH, size, null);
+    _insert(widget, DockLayoutConstant.NORTH, size, null);
   }
 
   /**
@@ -166,7 +164,7 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
    * @param size the child widget's size
    */
   void addSouth(Widget widget, double size) {
-    _insert(widget, LayoutDirection.SOUTH, size, null);
+    _insert(widget, DockLayoutConstant.SOUTH, size, null);
   }
 
   /**
@@ -185,7 +183,7 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
    * @param size the child widget's size
    */
   void addWest(Widget widget, double size) {
-    _insert(widget, LayoutDirection.WEST, size, null);
+    _insert(widget, DockLayoutConstant.WEST, size, null);
   }
 
   /**
@@ -207,7 +205,7 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
    *          <code>null</code> to append
    */
   void insertEast(Widget widget, double size, Widget before) {
-    _insert(widget, LayoutDirection.EAST, size, before);
+    _insert(widget, DockLayoutConstant.EAST, size, before);
   }
 
   /**
@@ -221,7 +219,7 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
    *          <code>null</code> to append
    */
   void insertLineEnd(Widget widget, double size, Widget before) {
-    _insert(widget, LayoutDirection.LINE_END, size, before);
+    _insert(widget, DockLayoutConstant.LINE_END, size, before);
   }
 
   /**
@@ -235,7 +233,7 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
    *          <code>null</code> to append
    */
   void insertLineStart(Widget widget, double size, Widget before) {
-    _insert(widget, LayoutDirection.LINE_START, size, before);
+    _insert(widget, DockLayoutConstant.LINE_START, size, before);
   }
 
   /**
@@ -248,7 +246,7 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
    *          <code>null</code> to append
    */
   void insertNorth(Widget widget, double size, Widget before) {
-    _insert(widget, LayoutDirection.NORTH, size, before);
+    _insert(widget, DockLayoutConstant.NORTH, size, before);
   }
 
   /**
@@ -261,7 +259,7 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
    *          <code>null</code> to append
    */
   void insertSouth(Widget widget, double size, Widget before) {
-    _insert(widget, LayoutDirection.SOUTH, size, before);
+    _insert(widget, DockLayoutConstant.SOUTH, size, before);
   }
 
   /**
@@ -274,7 +272,7 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
    *          <code>null</code> to append
    */
   void insertWest(Widget widget, double size, Widget before) {
-    _insert(widget, LayoutDirection.WEST, size, before);
+    _insert(widget, DockLayoutConstant.WEST, size, before);
   }
 
   /**
@@ -287,14 +285,14 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
    * @param before the widget before which to insert the new child, or
    *          <code>null</code> to append
    */
-  void _insert(Widget widget, LayoutDirection direction, double size, Widget before) {
+  void _insert(Widget widget, DockLayoutConstant direction, double size, Widget before) {
     assertIsChild(before);
 
     // Validation.
     if (before == null) {
       assert (_center == null); // : "No widget may be added after the CENTER widget";
     } else {
-      assert (direction != LayoutDirection.CENTER); // : "A CENTER widget must always be added last";
+      assert (direction != DockLayoutConstant.CENTER); // : "A CENTER widget must always be added last";
     }
 
     // Detach new child.
@@ -309,7 +307,7 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
       children.insert(widget, index);
     }
 
-    if (direction == LayoutDirection.CENTER) {
+    if (direction == DockLayoutConstant.CENTER) {
       _center = widget;
     }
 
@@ -363,7 +361,7 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
    *         child of this panel
    * @throws AssertionError if the widget is not a child and assertions are enabled
    */
-  LayoutDirection getWidgetDirection(Widget child) {
+  DockLayoutConstant getWidgetDirection(Widget child) {
     assertIsChild(child);
     if (child.getParent() != this) {
       return null;
@@ -443,7 +441,7 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
     assertIsChild(widget);
     LayoutData data = widget.getLayoutData() as LayoutData;
 
-    assert (data.direction != LayoutDirection.CENTER); // : "The size of the center widget can not be updated.";
+    assert (data.direction != DockLayoutConstant.CENTER); // : "The size of the center widget can not be updated.";
 
     data.size = size;
 
@@ -465,22 +463,20 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
 
   /**
    * Resolve the specified direction based on the current locale. If the
-   * direction is {@link LayoutDirection#LINE_START} or {@link LayoutDirection#LINE_END},
-   * the return value will be one of {@link LayoutDirection#EAST} or
-   * {@link LayoutDirection#WEST} depending on the RTL mode of the locale. For all
+   * direction is {@link DockLayoutConstant#LINE_START} or {@link DockLayoutConstant#LINE_END},
+   * the return value will be one of {@link DockLayoutConstant#EAST} or
+   * {@link DockLayoutConstant#WEST} depending on the RTL mode of the locale. For all
    * other directions, the specified value is returned.
    *
    * @param direction the specified direction
    * @return the locale
    */
-  LayoutDirection getResolvedDirection(LayoutDirection direction) {
-//    if (direction == LayoutDirection.LINE_START) {
-//      return LocaleInfo.getCurrentLocale().isRTL()
-//          ? LayoutDirection.EAST : LayoutDirection.WEST;
-//    } else if (direction == LayoutDirection.LINE_END) {
-//      return LocaleInfo.getCurrentLocale().isRTL()
-//          ? LayoutDirection.WEST : LayoutDirection.EAST;
-//    }
+  DockLayoutConstant getResolvedDirection(DockLayoutConstant direction) {
+    if (direction == DockLayoutConstant.LINE_START) {
+      return LocaleInfo.getCurrentLocale().isRTL() ? DockLayoutConstant.EAST : DockLayoutConstant.WEST;
+    } else if (direction == DockLayoutConstant.LINE_END) {
+      return LocaleInfo.getCurrentLocale().isRTL() ? DockLayoutConstant.WEST : DockLayoutConstant.EAST;
+    }
     return direction;
   }
 
@@ -524,31 +520,31 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
       }
 
       switch (getResolvedDirection(data.direction)) {
-        case LayoutDirection.NORTH:
+        case DockLayoutConstant.NORTH:
           layer.setLeftRight(left, _unit, right, _unit);
           layer.setTopHeight(top, _unit, data.size, _unit);
           top += data.size;
           break;
 
-        case LayoutDirection.SOUTH:
+        case DockLayoutConstant.SOUTH:
           layer.setLeftRight(left, _unit, right, _unit);
           layer.setBottomHeight(bottom, _unit, data.size, _unit);
           bottom += data.size;
           break;
 
-        case LayoutDirection.WEST:
+        case DockLayoutConstant.WEST:
           layer.setTopBottom(top, _unit, bottom, _unit);
           layer.setLeftWidth(left, _unit, data.size, _unit);
           left += data.size;
           break;
 
-        case LayoutDirection.EAST:
+        case DockLayoutConstant.EAST:
           layer.setTopBottom(top, _unit, bottom, _unit);
           layer.setRightWidth(right, _unit, data.size, _unit);
           right += data.size;
           break;
 
-        case LayoutDirection.CENTER:
+        case DockLayoutConstant.CENTER:
           layer.setLeftRight(left, _unit, right, _unit);
           layer.setTopBottom(top, _unit, bottom, _unit);
           break;
@@ -564,44 +560,23 @@ class DockLayoutPanel extends ComplexPanel implements AnimatedLayout, RequiresRe
 }
 
 /**
- * Used in {@link DockLayoutPanel#addEast(Widget, double)} et al to specify
- * the direction in which a child widget will be added.
- */
-class LayoutDirection<int> extends Enum<int> {
-
-  const LayoutDirection(int type) : super (type);
-
-  static const LayoutDirection NORTH = const LayoutDirection(0);
-  static const LayoutDirection EAST = const LayoutDirection(1);
-  static const LayoutDirection SOUTH = const LayoutDirection(2);
-  static const LayoutDirection WEST = const LayoutDirection(3);
-  static const LayoutDirection CENTER = const LayoutDirection(4);
-  static const LayoutDirection LINE_START = const LayoutDirection(5);
-  static const LayoutDirection LINE_END = const LayoutDirection(6);
-}
-
-/**
  * Layout data associated with each widget.
  */
 class LayoutData {
-  LayoutDirection direction;
+  DockLayoutConstant direction;
   double oldSize, size;
   double originalSize;
   bool hidden = false;
   Layer layer;
 
-  LayoutData(LayoutDirection direction, double size, Layer layer) {
-    this.direction = direction;
-    this.size = size;
-    this.layer = layer;
-  }
+  LayoutData(this.direction, this.size, this.layer);
 }
 
-class DockAnimateCommand extends LayoutCommand {
+class _DockAnimateCommand extends LayoutCommand {
 
   DockLayoutPanel _panel;
 
-  DockAnimateCommand(this._panel, Layout layout) : super(layout);
+  _DockAnimateCommand(this._panel, Layout layout) : super(layout);
 
   void doBeforeLayout() {
     _panel._doLayout();